requirements.txt
pip
Django~=5.1.6
Pillow
python-dotenv~=1.0.1
python-telegram-bot~=21.10